+++ sdk/resourcemanager/compute/armcompute/models.go
@@ -3961,6 +3961,21 @@ type ProximityPlacementGroupUpdate struct {
 	Tags map[string]*string
 }
 
+// ProxyAgentSettings - Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2023-09-01.
+type ProxyAgentSettings struct {
+	// Specifies whether ProxyAgent feature should be enabled on the virtual machine or virtual machine scale set.
+	Enabled *bool
+
+	// Increase the value of this property allows user to reset the key used for securing communication channel between guest
+	// and host.
+	KeyIncarnationID *int32
+
+	// Specifies the mode that ProxyAgent will execute on if the feature is enabled. ProxyAgent will start to audit or monitor
+	// but not enforce access control over requests to host endpoints in Audit mode,
+	// while in Enforce mode it will enforce access control. The default value is Enforce mode.
+	Mode *Mode
+}
+
 // ProxyOnlyResource - The ProxyOnly Resource model definition.
 type ProxyOnlyResource struct {
 	// READ-ONLY; Resource Id
@@ -4980,6 +4995,9 @@ type SecurityProfile struct {
 	// including Resource/Temp disk at host itself. The default behavior is: The Encryption at host will be disabled unless this
 	// property is set to true for the resource.
 	EncryptionAtHost *bool
+
+	// Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2023-09-01.
+	ProxyAgentSettings *ProxyAgentSettings
 
 	// Specifies the SecurityType of the virtual machine. It has to be set to any specified value to enable UefiSettings. The
 	// default behavior is: UefiSettings will not be enabled unless this property is